# docs/swagger.yaml
definitions:
  models.Clip:
    properties:
      creator:
        type: integer
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      game:
        type: integer
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
      viewcount:
        type: integer
      vod:
        type: string
    required:
    - date
    - duration
    - filename
    - resolution
    - size
    - title
    - viewcount
    type: object
  models.Creator:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: integer
    required:
    - name
    type: object
  models.Game:
    properties:
      box_art:
        type: string
      name:
        type: string
      uuid:
        type: integer
    required:
    - box_art
    - name
    type: object
  models.Vod:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      fps:
        type: number
      publish:
        type: boolean
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
    required:
    - date
    - duration
    - filename
    - fps
    - publish
    - resolution
    - size
    - title
    type: object
info:
  contact: {}
paths:
  /clips/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a clip
        in: query
        name: uuid
        type: string
      - description: The title of a clip
        in: query
        name: title
        type: string
      - description: The duration of a clip
        in: query
        name: duration
        type: integer
      - description: The date of a clip
        in: query
        name: date
        type: string
      - description: The filename of a clip
        in: query
        name: filename
        type: string
      - description: The resolution of a clip
        in: query
        name: resolution
        type: string
      - description: The size of a clip
        in: query
        name: size
        type: integer
      - description: The viewcount of a clip
        in: query
        name: viewcount
        type: integer
      - description: The creator id of a clip
        in: query
        name: creator
        type: integer
      - description: The game id of a clip
        in: query
        name: game
        type: integer
      - description: The vod id of a clip
        in: query
        name: vod
        type: string
      - description: 'Set order direction divided by comma. Possible ordering values:
          ''date'', ''duration'', ''size''. Possible directions: ''asc'', ''desc''.
          Example: ''date,desc'''
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Clip'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all clips
      tags:
      - Clips
    post:
      consumes:
      - application/json
      parameters:
      - description: Clip obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Clip'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Create clip
      tags:
      - Clips
  /clips/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete clip
      tags:
      - Clips
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Clip'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get clips by uuid
      tags:
      - Clips
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      - description: Clip obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Clip'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Patch clip
      tags:
      - Clips
  /creators/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a creator
        in: query
        name: uuid
        type: integer
      - description: The name of a creator
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Creator'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all creators
      tags:
      - Creators
    post:
      consumes:
      - application/json
      parameters:
      - description: Creator obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Creator'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Create creator
      tags:
      - Creators
  /creators/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete creator
      tags:
      - Creators
    get:
      parameters:
      - description: Unique Identifyer
        in: path
        name: uuid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Creator'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get creator by uuid
      tags:
      - Creators
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: integer
      - description: Creator obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Creator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Patch creator
      tags:
      - Creators
  /games/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a game
        in: query
        name: uuid
        type: integer
      - description: The name of a game
        in: query
        name: name
        type: string
      - description: The box_art of a game
        in: query
        name: box_art
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all games
      tags:
      - Games
    post:
      consumes:
      - application/json
      parameters:
      - description: Game obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Create game
      tags:
      - Games
  /games/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete game
      tags:
      - Games
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get game by uuid
      tags:
      - Games
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: integer
      - description: Game obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Patch game
      tags:
      - Games
  /vods/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a vod
        in: query
        name: uuid
        type: string
      - description: The title of a vod
        in: query
        name: title
        type: string
      - description: The duration of a vod
        in: query
        name: duration
        type: integer
      - description: The date of a vod
        in: query
        name: date
        type: string
      - description: The filename of a vod
        in: query
        name: filename
        type: string
      - description: The resolution of a vod
        in: query
        name: resolution
        type: string
      - description: The fps of a vod
        in: query
        name: fps
        type: integer
      - description: The size of a vod
        in: query
        name: size
        type: integer
      - description: 'Set order direction divided by comma. Possible ordering values:
          ''date'', ''duration'', ''size''. Possible directions: ''asc'', ''desc''.
          Example: ''date,desc'''
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vod'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all vods
      tags:
      - Vods
    post:
      consumes:
      - application/json
      parameters:
      - description: Vod obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Vod'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Create vod
      tags:
      - Vods
  /vods/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete vod
      tags:
      - Vods
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vod'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get vod by uuid
      tags:
      - Vods
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      - description: Vod obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Vod'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Patch vod
      tags:
      - Vods
swagger: "2.0"